Fix import of hostmaster item in SOA record
The import was correcting the email address (changing the first dot to an alias and removing the last dot). But in the zone edit entry, only domain format is enabled, not email format. An alternative fix would be to change it to 'email' type, but that would require some SQL update command when migrating to the new version. Additionally, it is common knowledge that email in an SOA record is written in domain format.